+++ src/main/java/com/example/trainerintake/dto/WorkoutPlanResponse.java
@@ -3,17 +3,17 @@ package com.example.trainerintake.dto;
 import java.util.List;
 
 public class WorkoutPlanResponse {
-    private Integer surveyId;
+    private Long surveyId;
     private List<String> answers;
     private String workoutPlan;
 
-    public WorkoutPlanResponse(Integer surveyId, List<String> answers, String workoutPlan) {
+    public WorkoutPlanResponse(Long surveyId, List<String> answers, String workoutPlan) {
         this.surveyId = surveyId;
         this.answers = answers;
         this.workoutPlan = workoutPlan;
     }
 
-    public Integer getSurveyId() { return surveyId; }
+    public Long getSurveyId() { return surveyId; }
     public List<String> getAnswers() { return answers; }
     public String getWorkoutPlan() { return workoutPlan; }
 }
